% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plots.R
\name{color_plots}
\alias{color_plots}
\title{Plot Color Palettes}
\usage{
color_plots(df, .group)
}
\arguments{
\item{df}{A data frame with the characteristics outlined in the
\code{description} section of the documentation.}

\item{.group}{A name describing a brand or project associated with the color
palette to be plotted.}
}
\value{
A ggplot2 plot.
}
\description{
This function helps visualize color palettes like the EWB brand
standard color palette created in the \code{EWBTemplates} package.

This function expects to be passed a data frame containing color palette
information. It expects the data frame to have at least one row and at least
4 columns. It expects those columns to include:

\describe{
\item{group}{
A name describing a brand or project associated with the color palette
(e.g., "ewb").
}
\item{subgroup}{
Optional names given to describe color categories (e.g., primary,
secondary, accent).
}
\item{hex}{Hexidecimal code for each color.}
\item{description}{
A descriptive name for each color (e.g., black, blue, dark green).
}
}
}
\examples{
data(ewb_colors)
color_plots(ewb_colors, "ewb")
}
