% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_cst_blanks.R
\name{cst_blanks_template}
\alias{cst_blanks_template}
\title{Return the "blanks replaced" CST template as a string}
\usage{
cst_blanks_template()
}
\value{
A single character string containing the CST template.
}
\description{
Returns a testthat block that fails if any \verb{____} remain in
\code{learner_code}. This function does not interact with RStudio.
}
