% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ewb_colors}
\alias{ewb_colors}
\title{Data frame containing Epi-Workbench brand standard colors}
\format{
A data frame with 5 rows and 4 columns:

\describe{
\item{group}{
A name describing a brand or project associated with the color palette
(e.g., "ewb").
}
\item{subgroup}{
Optional names given to describe color categories (e.g., primary,
secondary, accent).
}
\item{hex}{Hexidecimal code for each color.}
\item{description}{
A descriptive name for each color (e.g., black, blue, dark green).
}
}
}
\usage{
ewb_colors
}
\description{
A data frame containing Epi-Workbench brand standard color palette information.
}
\keyword{datasets}
