% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_cst_blanks.R
\name{insert_cst_blanks}
\alias{insert_cst_blanks}
\title{Insert the "blanks replaced" CST template at the cursor (RStudio addin)}
\usage{
insert_cst_blanks()
}
\value{
Invisibly returns the inserted template string.
}
\description{
Inserts the output of \code{\link[=cst_blanks_template]{cst_blanks_template()}} into the active
RStudio document at the current cursor position.
}
