% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hints.R
\name{insert_hints}
\alias{insert_hints}
\title{Insert Code Block Hints into a Quarto Lesson}
\usage{
insert_hints()
}
\value{
Invisibly returns the inserted hint template text.
}
\description{
Launches an RStudio Addin that interactively inserts a templated hint text
code block into the active Quarto lesson at the current cursor location.

This tool supports multiple predefined hint templates, including:
\itemize{
\item "default". The default hint template. This template includes general guidance for writing hints.
\item "no_mod". A hint template for code blocks that are intended to be run without modification.
}

\strong{Note:} This Addin must be run inside the RStudio IDE.
}
\section{Usage}{

Run interactively via the RStudio Addins menu or call \code{insert_hints()} in the console.
}

\examples{
if (interactive()) insert_hints()
}
\seealso{
\href{https://github.com/epi-workbench/EWBTemplates/wiki/Hints}{Hints wiki page}

Other Insert Addins: 
\code{\link{insert_coding_exercise}()},
\code{\link{insert_hidden_code_block}()},
\code{\link{insert_initial_code_block}()},
\code{\link{insert_lesson_yaml_header}()},
\code{\link{insert_noeval_code_block}()},
\code{\link{insert_solution_code_block}()},
\code{\link{insert_test_code_block}()}
}
\concept{Insert Addins}
