% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_quarto_lesson_coding.R
\name{new_quarto_lesson_coding}
\alias{new_quarto_lesson_coding}
\title{Create a Blank Quarto File - Coding Lesson}
\usage{
new_quarto_lesson_coding(
  file_name = NULL,
  file_loc = here::here(),
  template_name = "lesson_coding",
  interactive = FALSE
)
}
\arguments{
\item{file_name}{What do you want to name the new Quarto file you're creating
from the template Quarto file? If a name isn't provided, then the template
name will be used by default.}

\item{file_loc}{Where do you want the new template Quarto file to be stored?
The default location is the value returned by \code{here::here()}.}

\item{template_name}{The name of the Quarto skeleton file = "lesson_coding". You
should only have change this value if you change the name of the skeleton
file.}

\item{interactive}{If TRUE, an interactive text editor will pop up on screen
allowing users to make edits to the new Quarto file before saving it. This
accomplished via a call to \code{utils::file.edit}.}
}
\value{
A message to the console.
}
\description{
Create a Blank Quarto File - Coding Lesson
}
\examples{

# With defaults
# new_quarto_lesson_coding()

# Set a new file name
# new_quarto_lesson_coding("test_qmd_file")

# Set new file location other than here::here()
# No forward slash at the end
# new_quarto_lesson_coding(file_loc = "/Users/bradcannell/Desktop")
}
